refactor(app): :zap: Refactored GUI mode Session Status and Session Edit logic (simplified logic and code cleanup)
