+++ cmd/doc.go
@@ -1,9 +1,14 @@
-// Package main contains the application's entry points and command-line logic
+// Package main serves as the entry point for the BLE Sync Cycle (BSC) application
 //
-// BLE Sync Cycle (BSC) is a Go application designed to synchronize video playback with real-time
+// BLE Sync Cycle (BSC) is an application designed to synchronize video playback with real-time
 // cycling data from Bluetooth Low Energy (BLE) devices, such as cycling speed and cadence (CSC)
-// sensors. The cmd package provides the executable commands that launch and configure the main
-// features of the application.
-// Note that while this package serves as a bootstrap for both cli and gui packages, the actual
-// implementation details for GUI mode are contained within the ui and session packages.
+// sensors
+//
+// This package is responsible for:
+//   - Initializing the application's core services and logging
+//   - Parsing command-line flags to determine the operating mode (CLI vs GUI)
+//   - Bootstrapping the Session Manager for CLI mode or launching the GTK4 user interface
+//
+// While main handles the initial startup, the actual business logic is delegated to the
+// internal/session package, and the graphical interface is managed by the ui package
 package main
