+++ internal/config/doc.go
@@ -1,5 +1,11 @@
-// Package config manages application configuration for BLE Sync Cycle (BSC)
+// Package config defines the configuration structures and defaults for the application
 //
-// It provides structures, defaults, and utilities for loading and validating settings from a
-// configuration file (e.g., config.toml) or command-line flags.
+// It handles loading Settings from TOML files and validating them against application requirements
+// The configuration is split into logical sections:
+//   - AppConfig: general application settings
+//   - SessionConfig: specific settings for the current workout session
+//   - BLEConfig: Bluetooth sensor parameters
+//   - VideoConfig: video playback specific settings
+//   - SpeedConfig: wheel size and simulation settings
+//   - OSDConfig: On-Screen Display (OSD) customization
 package config
