+++ internal/logger/doc.go
@@ -1,5 +1,8 @@
-// Package logger provides logging utilities for BLE Sync Cycle (BSC)
+// Package logger provides a context-aware logging infrastructure
 //
-// It sets up log output, levels, and formatting for application diagnostics and debugging.
-// Logging levels include DEBUG (DBG), INFO (INF), WARN (WRN), ERROR (ERR), and FATAL (FTL).
+// # It supports leveled logging (DEBUG, INFO, WARN, ERROR, FATAL) and custom output formatting
+//
+// The logger is designed to be passed a context.Context, allowing for request-scoped or
+// operation-scoped logging context to be maintained (though currently primarily used for
+// cancellation propagation, it sets the stage for more structured logging)
 package logger
