+++ internal/services/doc.go
@@ -1,6 +1,6 @@
-// Package services implements core business logic and service orchestration for BLE Sync
-// Cycle (BSC)
+// Package services provides foundational service utilities for the application
 //
-// It coordinates the interaction between BLE devices, video playback, and cycling data
-// synchronization.
+// Services houses the ShutdownManager, which coordinates the graceful termination
+// of the application, ensuring that resources are cleaned up and goroutines exit
+// properly upon receiving system signals (e.g., SIGTERM, SIGINT) or internal errors
 package services
