+++ internal/video/doc.go
@@ -1,7 +1,8 @@
-// Package video provides video playback control with support for MPV and VLC media players.
+// Package video provides video playback control with support for MPV and VLC media players
+//
 // It manages playback speed synchronization with external speed controllers and provides on-screen
-// display (OSD) capabilities.
+// display (OSD) capabilities
 //
 // The package uses a mediaPlayer interface to abstract different player implementations, allowing
-// for easy extension to support additional future players.
+// for easy extension to support additional future players
 package video
